[Baekjoon 15828] Router 수정
